requirements.txt
numpy
pandas
pillow
matplotlib
